Added demonstrative tests for #43, moved admin tests into `test_admin.py`
