+++ simple_history/tests/admin.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 from django.contrib import admin
 
 from simple_history.admin import SimpleHistoryAdmin
-from .models import Poll, Choice, Person, Book
+from .models import Poll, Choice, Person, Book, Document, Paper
 
 
 class PersonAdmin(SimpleHistoryAdmin):
@@ -15,3 +15,5 @@ admin.site.register(Poll, SimpleHistoryAdmin)
 admin.site.register(Choice, SimpleHistoryAdmin)
 admin.site.register(Person, PersonAdmin)
 admin.site.register(Book, SimpleHistoryAdmin)
+admin.site.register(Document, SimpleHistoryAdmin)
+admin.site.register(Paper, SimpleHistoryAdmin)
