+++ simple_history/tests/tests/__init__.py
@@ -1,2 +1,3 @@
 from .test_models import *
+from .test_admin import *
 from .test_commands import *
